Finished the rudimentary implementation of Parser
Parser can now fully parse Strings to produce Items.

Should be refactored and there should be more comments, but it works at least.

Finished tryToExtractScore(), used it createItem() to create Item objects

Added a method to check if a string is composed of only whitespaces, and used it in createItem() to use the appropriate Item constructor.