Modified tryToExtractScore() method to let it extract negative scores
Modified the if conditions to add '-' chars.

Changed the String to a StringBuilder to make it cleaner

Main got some more stuff to allow it to perform a complete processing of input to output. I still need to do testing, (refactoring can wait) and write documentation, but SP 3 finally works